mittrock-spotify-5 Loading from Spotify ends when the rate limit is hit.
Added a wrapper around `spotify-web-api-ts` calls to catch **429** (Rate limit exceeded) errors, increase the time between requests, wait a bit, and retry. Relies on my branch of that library, which changes it to throw the raw Axios error instead of wrapping it, so that I can get to the actual error details.